% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{HNSC_GSE93157}
\alias{HNSC_GSE93157}
\title{HNSC_GSE93157}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Squamous head and neck patients in GSE93157 data set.
}
