% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{LGG_E_MTAB_6270}
\alias{LGG_E_MTAB_6270}
\title{LGG_E_MTAB_6270}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Brain Lower Grade Glioma patients in E-MTAB-6270 data set.
}
