% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{MEL_GSE91061}
\alias{MEL_GSE91061}
\title{MEL_GSE91061}
\format{
matrix
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Melanoma patients in GSE91061 data set.
}
