% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{MEL_GSE93157}
\alias{MEL_GSE93157}
\title{MEL_GSE93157}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Melanoma patients in GSE93157 data set.
}
