% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{MEL_Nathanson_2017}
\alias{MEL_Nathanson_2017}
\title{MEL_Nathanson_2017}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Melanoma patients in Nathanson_2017 data set.
}
