% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{MEL_phs000452}
\alias{MEL_phs000452}
\title{MEL_phs000452}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Melanoma patients in phs000452 data set.
}
