% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{MEL_PRJEB23709}
\alias{MEL_PRJEB23709}
\title{MEL_PRJEB23709}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Glioblastoma patients in PRJNA482620 data set.
}
