% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{nonsqNSCLC_GSE93157}
\alias{nonsqNSCLC_GSE93157}
\title{nonsqNSCLC_GSE93157}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Nonsmall-cell lung cancer patients in GSE93157 data set.
}
