% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{NSCLC_GSE135222}
\alias{NSCLC_GSE135222}
\title{NSCLC_GSE135222}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Nonsmall-cell lung cancer patients in GSE135222 data set.
}
