% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{RCC_Braun_2020}
\alias{RCC_Braun_2020}
\title{RCC_Braun_2020}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Renal cell carcinoma patients in Braun_2020 data set.
}
