% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{RCC_GSE67501}
\alias{RCC_GSE67501}
\title{RCC_GSE67501}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Renal cell carcinoma patients in GSE67501 data set.
}
