% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{data}
\name{STAD_PRJEB25780}
\alias{STAD_PRJEB25780}
\title{STAD_PRJEB25780}
\format{
S4
}
\source{
public data set
}
\description{
An SummarizedExperiment object contains expression matrix and clinical information of Stomach adenocarcinoma patients in PRJEB25780 data set.
}
